requirements.txt
Flask
gunicorn
PyPDF2
python-docx
langchain==0.2.12
langchain-core==0.2.29
langchain-community==0.2.11
langchain-google-genai==1.0.7
google-ai-generativelanguage==0.6.5
python-dotenv
pypdf
duckduckgo-search
docx2txt

